# docs/swagger.yaml
definitions:
  commons.SwaggerApiResponseError:
    properties:
      data: {}
      message:
        type: string
      success:
        default: false
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithoutData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
...
  models.BranchRequest:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple store management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple Store Management API
  version: "1.0"
paths:
  /api/branch:
    get:
      consumes:
      - application/json
      description: get all branch
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Branch
      tags:
      - branch
    post:
      consumes:
      - application/json
      description: create a new branch
      parameters:
      - description: Branch Request
        in: body
        name: branchRequest
        required: true
        schema:
          $ref: '#/definitions/models.BranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Branch
      tags:
      - branch
  /api/branch/{id}:
    delete:
      consumes:
      - application/json
      description: delete a branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Branch
      tags:
      - branch
    get:
      consumes:
      - application/json
      description: get branch by id
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch
      tags:
      - branch
    put:
      consumes:
      - application/json
      description: update a branch
      parameters:
      - description: Branch Request
        in: body
        name: branchRequest
        required: true
        schema:
          $ref: '#/definitions/models.BranchRequest'
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Branch
      tags:
      - branch
  /api/branch/detail/{id}:
    get:
      consumes:
      - application/json
      description: get branch detail
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch Detail
      tags:
      - branch
  /api/branch/employees/{id}:
    get:
      consumes:
      - application/json
      description: get branch with employees
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch With Employees
      tags:
      - branch
  /api/branch/items/{id}:
    get:
      consumes:
      - application/json
      description: get branch with Iitems
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch With Items
      tags:
      - branch
  /api/branch/top:
    get:
      consumes:
      - application/json
      description: get top branch
      parameters:
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Top Branch
      tags:
      - branch
...
  /api/salesData/branch/{branchId}:
    get:
      consumes:
      - application/json
      description: get sales data by branch
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Sales Data Branch
      tags:
      - salesData
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"